+++ src/main/java/lotto/configuration/Configuration.java
@@ -1,7 +1,6 @@
 package lotto.configuration;
 
-import component.Composer;
-import io.input.ConsoleWriter;
+import io.output.ConsoleWriter;
 import lotto.controller.LottoController;
 import lotto.domain.lotto.LottoGenerator;
 import lotto.domain.lotto.LottoStore;
@@ -20,11 +19,8 @@ public final class Configuration {
 
     public static LottoController createLottoController() {
 
-        final InputView inputView = new InputView();
-        final OutputView outputView = new OutputView(
-                new ConsoleWriter(),
-                new Composer()
-        );
+        final InputView inputView = new InputView(new ConsoleWriter());
+        final OutputView outputView = new OutputView(new ConsoleWriter());
         final LottoGenerator lottoGenerator = new RandomLottoGenerator();
         final LottoStore lottoStore = new LottoStore(lottoGenerator);
         final LottoRepository lottoRepository = new LottoRepository();
