+++ src/main/java/lotto/validator/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.validator;
 
 import java.util.regex.Pattern;
 import lotto.controller.exception.NumericInputValidationException;
@@ -21,7 +21,7 @@ public final class InputValidator {
     }
 
     /**
-     * 입력이 ,로 구분된 문자열인지 검증
+     * @apiNote 검증 시, <h3 color="#bf0f4d">input은 반드시 ,로 구분된, 숫자로 이루어진 문자열이여야 합니다.</h3>
      */
     public static void validateCommasSeparatedInput(final String input) {
         if (!Pattern.matches(COMMAS_SEPARATED_REGEX, input)) {
@@ -30,7 +30,7 @@ public final class InputValidator {
     }
 
     /**
-     * 입력이 숫자로만 이루어져 있는지 검증
+     * @apiNote 검증 시, <h3 color="#bf0f4d">input은 반드시 숫자로만 이루어진 문자열이여야 합니다.</h3>
      */
     public static void validateNumericInput(final String input) {
         final char[] chars = input.toCharArray();
